requirement.txt
certifi==2025.8.3
charset-normalizer==3.2.0
configparser==6.0.0
kafka-python==2.2.15
requests==2.32.4
urllib3==2.5.0
